Refactor code into MonkeyText
